+++ server/impl/src/main/java/com/walmartlabs/concord/server/events/OneOpsEventResource.java
@@ -28,6 +28,7 @@ import com.walmartlabs.concord.server.org.project.ProjectDao;
 import com.walmartlabs.concord.server.org.project.RepositoryDao;
 import com.walmartlabs.concord.server.org.triggers.TriggersDao;
 import com.walmartlabs.concord.server.process.ProcessManager;
+import com.walmartlabs.concord.server.process.ProcessSecurityContext;
 import com.walmartlabs.concord.server.sdk.ConcordApplicationException;
 import com.walmartlabs.concord.server.user.UserManager;
 import io.swagger.annotations.Api;
@@ -81,9 +82,10 @@ public class OneOpsEventResource extends AbstractEventResource implements Resour
                                ProjectDao projectDao,
                                RepositoryDao repositoryDao,
                                TriggersConfiguration triggersCfg,
-                               UserManager userManager) {
+                               UserManager userManager,
+                               ProcessSecurityContext processSecurityContext) {
 
-        super(cfg, processManager, triggersDao, projectDao, repositoryDao, triggersCfg, userManager);
+        super(cfg, processManager, triggersDao, projectDao, repositoryDao, triggersCfg, userManager, processSecurityContext);
         this.objectMapper = new ObjectMapper();
     }